Suppress bash/ls warning
We got an error from the shell expansion of a glob when there are no include
files. It's completely unnecessary to use a glob and in fact, the output
produced by this invocation of ls needs to be post-processed using basename. By
changing our invocation to just list the files in the directory, we completely
avoid that extra invocation of basename, speeding this whole thing up!