Add ability to deal with extra stuff in NSO version
The version-gen script previously used the NSO version, like 5.3.2 but in a
tuple form like (5, 3, 2) as the unique identifier of each version. We stored it
in a tuple form since it allows us to easily sort or filter on parts, like
filtering out major version 5.

Nightly build or other special builds pose a challenge as we could have multiple
releases with the same basic NSO version, for example:

  nso-5.4-20200215_ps.linux.x86_64.installer.bin
  nso-5.4.linux.x86_64.installer.bin

We have to keep track of 5.4-20200215_ps as separate from 5.4, yet we need
5.4-20200215_ps to be treated as '5.4' when it comes to sorting. This patch
achieves that by keeping track of the version number in both forms.