Specify the CI runner image
Specify the Docker image that the CI executor should use.

We used to not define this because we wanted the choice to be up to the
user/administrator of a mirror of this repo and not specifying the image in
.gitlab-ci.yml makes it possible to configure it via the CI runners config by
setting the default image. Things have changed and I now think this is the best
alternative as it provides a sensible default yet is changeable thanks to our
continuous mirroring approach that allows for local changes.

From the very beginning we did have an image set. It was debian:buster, which
would be periodically refreshed in the upstream repo - we didn't have to think
much about making sure it didn't go stale. Unfortunately the debian image didn't
have all the tools we wanted out of the box, so the first part of each CI run
consisted of installing those missing prerequisites.

The installation took around 3 minutes for every job and so it quickly gets
boring. I figured that if we removed the image completely, it would be up to the
user to pick if they wanted to use debian or build their own image, that
contains the necessary tools and thus speed up the build. The easiest way to
allow such configuration was to simple remove the image statement from
.gitlab-ci.yml and allow it to be set in the CI runner configuration. That does
require administrator privileges on the CI system though. This also worked well
with standard Gitlab mirroring as no changes needed to be done to the code.

Since then, we have built our own image -
registry.gitlab.com/nso-developer/ci-runner-image and it is rebuilt on a nightly
basis to pull in all the latest security patches. GitLab mirroring has been
replaced with our own integrated continuous mirror approach that allows for
local changes, and so the reason not to provide an image is gone and the best
image to pick is the ci-runner-image, thus we are switching to it!