Fix testenv-build and include build-info
The testenv-build target was simply broken. I'm really quite surprised how I
must have committed broken code when I know that I tested all the cases and did
it on the bgworker repo. Somehow I must have broken things when I transferred
that code to the upstream skeletons. Anyway, it is fixed now.

We now generate a build-meta-data.xml and inject into the package, if the
package does not already have one. Very useful for operations!