Include empty extra-files dir
This fixes the standard build, since we actually rely on this directory being
present. To track it in git, I've added a .gitkeep file and then added
extra-files/.gitkeep to .dockerignore so that it is not actually copied over
into our container image. I opted for .gitkeep instead of .gitignore since I
want to ignore it in .dockerignore and we might actually want a .gitignore file
of things in extra-files...