Simplify netsim startup script
This radically simplifies the netsim startup script. We hard code paths to the
netsim directory, dynamically discover the package name on startup so we don't
have to be fed in arguments and remove a bunch of functionality from the run
script. The result is something vastly easier to read and understand.

The is-alive call to ncs-netsim has been replaced with simply waiting on the
confd pid. This leads to much lower CPU utilization, which was previously an
issue.

Signal handling for INT (sent on ^C), SIGQUIT and SIGTERM are properly
implemented making it possible to shut down the confd container cleanly and
later starting it again.

Process control is properly implemented so that while running the netsim
container in the foreground, it is possible to issue ^C and have a clean shut
down.

The netsim container now supports restarting confd by sending SIGHUP to the
container. Since confd doesn't have 'request packages reload', this does the
equivalent and allows us to load new models while the container is running,
while also preserving its running state (CDB config etc).

The netsim directory (/nsim) is placed on an anonymous volume, making it
possible to mount up from another container. This is turn should make it
possible to modify the netsim environment, for example inserting new .fxs files,
to alter the simulated device while the container is running. confd do need to
be restarted but should start up again with the previously running state (CDB
config).

For some history and understanding of why this script was so convoluted; it is
inherited from the TeraStream environment where netsim are used to test software
upgrade. The test is however a very specific one and the implemented in the
run-netsim.sh script was specific to the test rather than generic functionality
for testing software upgrade. I didn't realize this at first, which is why the
script was brought in verbatim from TeraStream as I figured it was nice
functionality to have. Given how non-generic it was, I now opted to remove it.
In addition, it actually turned out to be broken as multiple people have
modified the file over time and accidentally removed some pieces of code.

It should be possible to simulate software upgrade through the combination of
inserting new .fxs files in the /nsim directory and restarting confd.