Revamp testenv-build
testenv-build did half the job, like copying over the source and recompiling the
packages, inside a container using the cisco-nso-dev image but did the other
half of the job - reloading packages - from outside the container. This was
because we couldn't interact with the running NSO from within the build
container. Running on the host and not in a container meant that we relied on
the tools present on the host rather than in our cisco-nso-dev image. For
example, we made use of xmlstarlet, which most people probably do not have
installed.

This change restructures how testenv-build is carried out. Everything is now
running from within a container and thus, we have no dependence on xmlstarlet
being installed on the host. There were also some incompatibilities with how sed
was used which wasn't POSIX / BSD / OS X compliant. I fixed that too by using
sed -E and a slightly different syntax, however it is sort of unnecessary as we
now run sed inside of a container, which means we use GNU sed. The container is
run with --network=container:${NSO} which means we are in the same network
namespace as the running NSO container, which allows us to access the NSO IPC
port and thus we can use ncs_cli from our cisco-nso-dev container. Nice!