Use container hostname for netsim device name
When using the ncs-netsim script to create a netsim device, it uses the provided
name for the prompt of the netsim. I have understood this is desirable.

We used to create the netsim with the name 'dev', which has the advantage that
it is located at a known path (/netsim/dev/dev) so if we want to make later
modifications, like injecting shell scripts to emulate show commands, that
becomes fairly easy.

This change takes a new approach where we first generate the netsim device based
on the hostname of the container, to get the prompt right, and then moves the
resulting directory to /netsim/dev/dev before running it.

Do note that docker containers per default don't have the hostname set to the
name of the container (that would have been nice). Rather, they get some
randomly generated string. --hostname has to be specified to docker run. We set
the hostname for the dev1 container we start in the NED skeleton.