Use expect for SSH tests
Passing a command to ssh so that it would run in the NSO CLI used to work, like
we could do `ssh my-ncs 'show ncs-state version'` and get the version back. In
NSO 5.4 this appears to have been broken and our SSH client will exit with
status code 255. While it's good that we noticed that with our tests, the
purpose of our tests is actually not to test that particular feature, rather we
want to make sure that the user creation mechanism works. Thus, we here rewrite
our tests to use expect to work around the bug. A separate ticket has been
opened for the bug itself.