Combine base and dev image build
Instead of having two separate directories, each to build the base and the dev
image respectively, we now have a single directory, so that we more easily can
share common files and a single Dockerfile with multiple stages.

The Dockerfile is a little tricky, due to how we need to build the images. The
first nso_install target merely installs NSO and is later used by the dev image,
since we want the complete NSO install (including docs etc). The nso_stripped
target is based on the nso_install target but strips aways docs and similar
things to get a smaller result, which is later used in the base image.

debian_base installs packages that are common to the base and dev image, from
which the base and dev images are later based on.

Base and dev both start from debian_base and then copy in NSO from nso_install
and nso_stripped respectively, plus also installing more packages in the dev
image.

This also allows us to get rid of the awkward need to synchronize sections of
the Dockerfiles between the base and dev image, since it is now placed in the
debian_base target.

run-nso.sh is now available in the dev image, making it possible to easily start
NSO in the same way as in the base image.