Describe how to pick skeletons, slicing etc
I've gotten the question on how to get started using the skeletons, like what to
place where etc. I added a short description of a good starting point, including
that most users probably do not need the package skeleton.

Added a section explaining the pros and cons of slicing and dicing packages in
different ways, as a guideline for those looking to do more complex project
setups.