Reverse order of commands to avoid docker bug
As reported in https://gitlab.com/nso-developer/nso-docker/-/issues/162, it
appears we are hitting a bug in Docker
(https://github.com/moby/moby/issues/37965) that leads to a build failure on the
first build. A workaround is to add a dummy command in between but I found that
simply swapping the two COPY commands also fixes this. Avoiding a dummy command
means we also avoid a dummy layer in the output image.

I've run numerous test runs (and naturally cleared the intermediate docker
images from the build) to make sure there are no intermittent failures going on.
It 100% works with this fix and is 100% reproducible without it.