Avoid git errors for initial build
I think this could be considered mostly a cosmetic bug as an error doesn't abort
the build process and doesn't change the build output, it merely prints an error
in the build log. Nonetheless, that can be discerning for users and so we
suppress any such errors.

The situation arises when a user builds their project without having initialized
git, which is perhaps likely to happen the first time a user tries to build the
project, before they've committed the project to git. Since the project isn't
yet a git repository, running git commands to inspect the repo will lead to
failures. We now suppress those errors.