Source Python venv in NSO packages
This includes a custom Python VM startup script that attempts to detect a
Python virtualenv directory in a package and if found, activate that virtualenv
before starting the Python VM.

The path to the virtualenv is hardcoded relative to the package path:

  ${PKG_PATH}/pyvenv

NSO inserts the path to our package and places it in PYTHONPATH, however,
there could be other paths in the PYTHONPATH and so we have to parse out ours.
Similarly, we don't specifically know which package we are running for, so we
find out by digging through the arguments we pass along. -i is the "VM id",
which is the same as the package name. Through some parsing of the command line
arguments we find it. Hopefully this is more robust than expecting it in a fixed
position (it is currently $6).

The advantage of placing a venv in an NSO package is that the NSO package can be
truly self-contained. Otherwise we would depend on system installed Python
libraries. With Python libraries on a global system level, it becomes impossible
to support multiple different version of the same package, which is generally
desirable. NSO packages with Java are already self-contained and this is very
useful, for example NEDs that include NEDCOM can include different versions of
NEDCOM without collision / non-resolvable dependencies.

Placing the virtualenv in the ${PKG_PATH}/python/venv was considered, but as
${PKG_PATH}/python is placed in PYTHONPATH, placing anything there that is not a
Python package runs the risk of conflicting with an actual Python package. Thus,
placing it outside of the python directory appears cleaner.