Add test of venv
We test that the Python virtualenv loading works by creating a new package that
depends on package 'foo' that is installed in the virtualenv. If the activation
of the venv doesn't work, we will see a package load error in NSO.

The python venv itself is a minimal one, using --symlinks, so the size is ~40KB
instead of ~7MB when including the python and pip binaries.

I learned that venvs are not movable as they use absolute paths. This means that
when creating the venv you must do so with the package at its final location,
i.e. in /var/opt/ncs/packages. I did this by starting a dev container with the
package mounted in the right place, i.e.:

  cd test/packages/
  docker run -it --rm -v $(pwd):/var/opt/ncs/packages cisco-nso-dev:5.3.2-kll
  # in container
  cd /var/opt/ncs/packages/testpkg-pyvenv
  python3 -m venv --symlink --without-pip pyvenv
  touch pyvenv/lib/python3.7/site-packages/foo.py

I wonder if this will cause us problems during linting runs and similar!? Like
if we import some package, then pylint will complain if it can't find it - it
has to load the venv, but it is run during build time when the package lives in
/src/packages IIRC. Well, that's a separate issue...