Include venv in -dev image
We want to be able to create virtualenvs and since virtualenvs use absolute
paths and can't be moved, it has to be created in what will be the final
location of the package, i.e. /var/opt/ncs/packages/... This is quite easy to do
from within a container, since we can mount up the directory as
/var/opt/ncs/packages. The python venv module is now included in the -dev image,
so we can use it to create the virtualenv.