Add script to build venv from requirements.txt
One would imagine this to be a one-liner but it's a little more involved than
that. We only build a venv if a requirements.txt file exists in src. We create
the venv with symlinks to reduce the size and use system site package to bring
in pylint (without also having to install it (again) in our venv).

Since venv uses absolute paths and resolves its own location first, we need to
actively undo that and rewrite the path using sed.

We then install the packages specified in requirements.txt but ignore the system
installed packages, since we want to be able to get newer versions as well as
force installation of packages. The system installed packages differ between the
-dev image and the -base image and so if we didn't do this, the installer might
assume a package is installed (chardet for example), since it exists in the -dev
image, but when run in the -base image, it would be lacking.