Revamp build process to honor venv & fix pylint
We have to fake an environment that looks like the final installed location of
our packages for pylint to work properly. This allows us to reach dependencies
where we expect them. In addition, now that we support virtualenv, we must load
that virtualenv before building our packages. It's a bit of a dance but still
readable IMHO ;)