Perfect the illusion
TIL that entering a symbolic link directory is an illusion held up by bash and
friends. Let's say we have directory /a and /b is a symlink to /b. Doing `cd /b;
pwd` in bash will print '/b' as the current working directory. Bash is lying to
us. Let's say there is a Makefile in /a. Doing `cd /b; make` will have /b as the
current directory whereas doing `make -C /b` will have /a as the current
directory. It's easy to think of this as some extra de-referencing is happening,
but this is actually the normal behavior chdir.

We have to adapt to this reality and thus add extra symlinks to perfect the
illusion.