Improve handling of special builds
The tip of train (tot) version set is easy to understand for the simple case of
standard builds. For example, given these version:
- 5.3.1
- 5.3.2
- 5.4
- 5.4.1

As TOT are the latest maintenance releases for all major/minor combinations, the
TOT set includes:
- 5.3.2
- 5.4.1

However, what happens when we have a specialized build? Like a nightly release
or something build from a branch, like this:
- 5.3.1
- 5.3.2
- 5.3.2_ps
- 5.4
- 5.4.1

This would historically produce:
- 5.3.2_ps
- 5.4.1

While I think the correct, or at least better, output would be:
- 5.3.2
- 5.3.2_ps
- 5.4.1

That is, the presence of a special build should not overshadow the non-special
build of the same version. Similarly, if we have:
- 5.3.2
- 5.3.2_ps
- 5.3.2_nightly_something

We want the output to include all three. This commits achieves that.

I've also added in the "extra" part into the NsoV tuple. Not that we need it
right now, but I imagine it can be useful at some point.

A harder question to answer is what the output should be for:
- 5.3
- 5.3_ps
- 5.3.2

Should 5.3_ps be included in the output, since it is the last version if we
consider 5.3 with the "ps" patch to be a major / minor combo... or should be
major/minor combo strictly look at the numeric versions, thus not include 5.3_ps
in the output?