Force regen for version-set CI job
We have to force this make target to run to get predictable results. Make will
per default only inspect the modification timestamps of the output and
input (target and prerequisites respectively). Since both input and output are
checked into git, the timestamps will be bogus (when we checked out the files)
rather then reflect the true modification times of the files. Thus, we have to
force make to run the target to be sure this check is valid.

In addition, if we change the version-gen program, as we did in the previous
commit, that is not caught by the make target either.

As the time it takes to run the target is negligible, there is no real downside
here.