Pull latest images for build
Using --pull to docker build will ensure we always grab the latest version of
whatever other Docker images we depend on, for example the NSO image itself. If
we are running on a machine that has the image, say
registry.example.com/nso-docker/cisco-nso-dev:5.4, that version will be used,
even if a newer image is available in the registry. Using --pull will ensure we
check with the registry and pull down the latest version before building.

As the base nso-docker images get new features and functionality, we want to
make sure we have the latest or the builds might fail in various ways.

This doesn't show up in CI when the CI runners are single-use, as they are for
gitlab.com/nso-developer/ and that's why this hasn't been unearthed earlier.