Correct RELOAD_PATTERN
The entirety of the testenv-build function was previously defined in a make
target but as it grew I found it easier to split out into its own script. In
make, $ needs to be escaped as $$. However, as the functionality was separated
into a script, this wasn't done and we kept the escaping as done in makefile,
resulting in an incorrect regexp used here ($$ will evaluate to the current PID
instead, thus the regexp won't match on the right thing).