Extend testenv-start for Python Remote Debugging support
Setting up Python Remote Debugging in VSCode requires the container to
expose a listening port (default 5678). In a multiuser environment the
port must be randomzied - we let docker do that with '-P'. The extra
target then finds the assigned host port for the current environment and
updates VSCode launch.json if found and if it contains the remote
debugging target.