Verify packages are operationally up in skeletons
We don't ship much tests in the skeletons per default but this seems like fairly
non-controversial to add. Making sure the package loads up in NSO is a good
baseline!

We use negative lookahead in our regexp 'oper-status (?!up)' to match a line
with 'oper-status ' but not with 'up' at the end, i.e. all other states. Then
inverse the return code through the '!' operator so that finding such a line
constitutes an error (normally grep expects to find the line, but such a line
is *bad* in our case). If an error is encountered we run "show packages" to
provide more debug information to the user.

GNU grep supports negative lookahead when in PCRE mode (-P). OS X does not come
with GNU grep per default and so we run grep inside of the cisco-nso-dev
container in order to get a GNU grep. A little convoluted but it works.