Add docker build cache argument to NID skeletons
We can now control whether the cache is used in the docker build process in NID
skeletons. Per default we disable the cache when building in CI on the default
branch (typically 'master'). We previously had the same check (i.e. disabling
cache) on the main nso-docker build but had hard-coded the branch check to
'master'. It is now using the CI_DEFAULT_BRANCH variable instead. Much better!