Create a separate python VM startup script for remote debugging
Just enabling remote debugging based on the presence of the debugpy
package is a security risk. If the package is deployed, any user
connected to the docker host where the NSO container runs could
potentially access the listening port. If we use a separate Python VM
startup script, the decision to run the debugger is explicitly done at
runtime. ncs.conf (startup config) has the option
/ncs-config/python-vm/start-command.

The script ncs-start-python-vm-debug is a symlink to the original
ncs-start-python-vm. By checking the name of the executable it takes a
decision to include debugpy at runtime.