Ensure we use latest NID images
This changes the check-nid-available target into ensure-fresh-nid-available and
as the name suggests it not only checks that we got a NID image, we also attempt
pulling in the latest versions.

It is possible to skip pulling by setting the environment variable SKIP_PULL to
true. Useful when you're on an airplane or similar.