Add target to wait on NSO startup
tesenv-wait-started-nso will wait, up to 600 seconds, for all NSO instances to
start up. It uses `ncs --wait-started` to detect if startup is successful or
not.

It's 600 seconds per NSO instances, which means the first in the list we check
will have 600 seconds whereas the second one actually gets closer to 600+600 and
so forth.