Include less in the -base image too
'less' is very useful for perusing the logs inside of a running
container. It is already included in the -dev image. It would also be
possible to use the -dev image to mount the log volume from -base. But
that requires foresight - creating the NSO container with the /log
volume.