Add test packages for inter-package pyvenv dep test
testpkg-pyvenv-a depends on foo that is in the virtualenv of testpkg-pyvenv-a,
so if that imports correctly we can detect that the local package virtualenv is
correctly loaded. testpkg-pyvenv-a also depends on testpkg-pyvenv-b, which in
turn depends on package bar, which is in the virtualenv of testpkg-pyvenv-b.
Thus, for testpkg-pyvenv-a to correctly load, we must support adding the
virtualenv of another package to our path.

testpkg-pyvenv-b is equivalent to testpkg-pyvenv, just renamed.