Load venv of dependencies
We have previously supported loading a Python virtualenv when starting the
Python VM for a particular NSO package. This makes it possible to ship an
entirely self-contained NSO package, which is great. However, NSO also makes it
possible to define dependencies between NSO package, i.a. NSO package A can
depend on NSO package B and when starting A, the PythonVM will be fed with a
PYTHONPATH that includes the python/ directory of package B, making it possible
to import packages from B. However, if B also comes with a virtualenv and have
its own dependencies installed in that venv, they will not be visible from the
PythonVM for A. This patch addresses that.

It is only possible to activate one virtualenv at a time so in order to bring in
the packages installed in the virtualenv of other NSO packages, we explicitly
add the site-packages from the virtualenv to PYTHONPATH. The first virtualenv,
the virtualenv from our own package, is still activated rather than just added
to PYTHONPATH. There is an important distinction in that activating a virtualenv
will use the python binary in that virtualenv. It is thus possible to ship a
different version of Python, for example Python 3.8 instead of Python 3.7, in a
virtualenv. Note how this also means it is possible to have a version mismatch
between a package and other NSO packages it might depend on. Sourcing the
virtualenv from another package is not possible unless both packages use the
same major/minor Python version.

The changes are to the python vm startup script, which means this only takes
effect at run time. Developers of these NSO packages will also want to activate
and source the virtualens across multiple NSO packages in a similar way at
development time, so that linters and other code tools can be used. This is
outside the scope of this patch.

The path handling is somewhat simplistic, assuming no spaces, but I'm not sure
if spaces in paths would be supported anyway?