Create a DAG by declaring job dependencies for build -> test
The stages build, test and deploy contain many jobs for individual
versions. Because individual jobs target a specific NSO version (or in
the case of multiver, two versions), the pipeline can proceed to the
test stage for an NSO version without requiring all build jobs to pass.
This may also reduce the time it takes to run the pipeline, but the
primary driver is the ability to isolate failures in the build stage.

There is a reason dependencies are only done declared for the build ->
test transition and not for test -> deploy too. We want to guarantee
feature parity for the repository across all NSO versions tested. Only
until all test jobs have passed do the build jobs start. This is the
current behavior too.