Use expect for DAP positive test
The old test was apparently a little flaky - not in the CI build for the
authoritative repo but it is important that this runs well in customer
installations and thus that we spend a little time on making things tick.

We here replace a rather simplistic method of checking for debugpy with that of
expect which can be much more refined. In particular, we allow for a long total
timeout (10 seconds) but also allowing the test to finish as soon as possible
With the sleep, the test will take a minimum of the sleep time (1 second)
regardless of positive or negative outcome. The new test can finish faster than
that as we aren't passively sleeping - we are now actively waiting for certain
input.