Use DOCKER_TAG instead of PNS
In this code we are conceptually working with the docker tag but we manually
composed it each time of $(NSO_VERSION)-$(PNS) - which is exactly what
DOCKER_TAG consists of, though it allows overriding it. Since we are
conceptually working with the docker tag, we switch to the DOCKER_TAG variable
since it is meant exactly for this.