Use proper docker labels for containers
They're a bit long but I guess that's what happens when you follow the
hierarchical structure route...

We group together the name and type labels under the hirerarchiy
com.cisco.nso.testenv which clearly denotes this is for NSO (we don't need to
say NID / -in-Docker because we are working with containers, so that part should
be clear) and shows that it's related to testenvs specifically.
com.cisco.nso.testenv.name refers to the name of the testenv, which is shared
for all the containers in that testenv, whereas com.cisco.nso.testenv.type
refers to the type of a single container - is this confusing?