Support nightly and special builds in version-gen
This improves the tip-of-train (tot) computation for all known types of NSO
releases, i.e.;
- normal releases
- special builds, like for pushed-status or similar built from a branch
- nightly builds

To ensure the right behavior, a small test suite is added. Unfortunately due to
the naming of the version-gen file, we can't import it, so use a symlink to get
around that.

Computing tip of train for normal releases is fairly straight forward, like with
releases;
- 5.3
- 5.3.1
- 5.3.2
- 5.4
- 5.4.1

It is clear that we should only get 5.3.2 and 5.4.1 in the result.

Special builds should be, by way of best practice, include the branch/feature
name in the version, like 5.4_ps for the pushed-status feature. When we have a
version-set with a mix of normal and special builds, we want to preserve both,
i.e. with the following input:
- 5.4
- 5.4_ps
- 5.4.1

We want to get 5.4_ps and 5.4.1.

Nightly builds can look like:
- 5.4.2_200922.050200687.7da830a96f9c
- 5.4.2_200923.050200682.b84c2c6f442e

The date is included in the build name, so we just want to pick the latest
nightly build, but again, it should not exclude the normal build, nor that of a
special build.