Show package oper-status if redeploy fails
The last step of the testenv-build script is to redeploy the modified
packages. If the step fails ("result false") it is very convenient if we
immediatelly display "show packages package oper-status" to the user.
The "request packages reload" command already shows package load errors
in its output so this extra step is not necessary.