Add README in GitHub flavored MarkDown
Cisco DevNet Code Exchange is only able to parse Markdown files and has a
requirement that a README file, in Markdown format, be present.

I contemplated replacing the org-mode README with Markdown so that we wouldn't
have to maintain two files and risk them going out of sync. However, it's not
just that syntax of Markdown is worse (it is!), we'd also loose features. There
are org-babel expressions in README.org that can be executed to capture output
that are included in the file. There's no equivalent in Markdown as far as I
know, thus I have opted to let the org-mode README remain as the authoritative
master and have here exported in using Emacs to GitHub flavored Markdown. I hope
we can keep the files in sync manually.

I did ponder a CI job, but it would be quite an effort to add for quite a small
gain. I don't have all day ;)