Add IPv6 support
Docker cannot automatically assign an IPv6 prefix like it can with IPv4. We now
generate a random IPv6 prefix from the Unique Local Addresses (ULA) range and
use as our IPv6 network prefix. ULA can be used somewhat like RFC1918 IPv4
addresses though with no overlap and no NATing. However, they may not be used on
the public Internet meaning that we will have no external IPv6 connectivity. To
avoid this, we don't have a default route for IPv6, thus ensuring we only get
local connectivity. Docker doesn't actually allow us to not have a default
gateway - it will always add one, but we remove it after the containers have
been started to avoid broken connectivity. This is achieved by using a helper
container, that should make it possible to run this on Linux, OS X as well as
Windows (unlike some combination of nsenter which would be Linux only). The
default route will only be removed if the IPv6 network is not public.

It is possible to disable IPv6 by setting the environment variable IPV6=false.

It is possible to manually specify the prefix to use by setting IPV6_NET, for
example IPV6_NET=2001:db8:1234:5678: