Run test target if existent & fix strip
We introduce the ability to run a 'test' target in a package that we are
compiling. Before running it, we check if it exists.

The strip target is also reworked to follow the same concept as the test target.
We rely on if instead of && and a sub-shell, which is broken, as the  || exit
will only affect the sub-shell return code and thus have no effect on the outer
shell return code.

James originally submitted the code for the strip target. I blatantly rewrote it
and squashed the commits, in an attempt to give him the credit I rewrote his
commit!, but it all failed as my code was broken and James original code was
correct. I attempted to fix it by adding sub-shell execution but that failed in
a different way (strip failure wouldn't be propagated). We're now coming full
circle and are back at what James originally submitted (modulus some very small
changes). Thanks James Harr.