Run check-logs from main CI script
check-logs is now run in the main CI script in addition to the after_script.
This is done just after the tests but before push. Doing so means we will be
able to detect errors and error out from the CI job, with the CI job accurately
reflecting the error and aborting.

Running check-logs in the after_script does not have an effect on the CI job
status. That is, even if we see errors in the logs and we return a non-zero
return code, the CI job status will still succeed simply because after_script
does not influence the CI job status.

Additionally, in the after_script we change how this target is run, since we
don't actually want the after_script to terminate early (though this is
irrelevant on the return code mentioned above). After the check-logs target in
after_script we save the logs as artifacts and stop the testenv. We never want
to skip those tests. Thus, we now ignore any return code from check-logs in
after_script and continue regardless so we can guarantee we stop the testenv as
we should.